+++ api/tests/unit/integrations/launch_darkly/test_client.py
@@ -214,7 +214,7 @@ def test_launch_darkly_client__get_flag_tags__return_expected(
 
 @pytest.mark.parametrize(
     "response_headers",
-    [{"Retry-After": "0.1"}, {"X-Ratelimit-Reset": "1672531200.1"}],
+    [{"Retry-After": "0.1"}, {"X-Ratelimit-Reset": "1672531200100"}],
 )
 @pytest.mark.freeze_time("2023-01-01T00:00:00Z")
 def test_launch_darkly_client__rate_limit__expected_backoff(
@@ -252,7 +252,7 @@ def test_launch_darkly_client__rate_limit__expected_backoff(
 
 @pytest.mark.parametrize(
     "response_headers",
-    [{"Retry-After": "0.1"}, {"X-Ratelimit-Reset": "1672531200.1"}],
+    [{"Retry-After": "0.1"}, {"X-Ratelimit-Reset": "1672531200100"}],
 )
 @pytest.mark.freeze_time("2023-01-01T00:00:00Z")
 def test_launch_darkly_client__rate_limit_max_retries__raises_expected(
@@ -325,3 +325,39 @@ def test_launch_darkly_client__rate_limit_invalid_response__raises_expected(
     # When & Then
     with pytest.raises(HTTPError):
         client.get_project(project_key=project_key)
+
+
+def test_launch_darkly_client__rate_limit_no_headers__waits_expected(
+    request: pytest.FixtureRequest,
+    mocker: MockerFixture,
+    requests_mock: RequestsMockerFixture,
+) -> None:
+    # Given
+    token = "test-token"
+    project_key = "test-project-key"
+
+    mocker.patch(
+        "integrations.launch_darkly.client.BACKOFF_DEFAULT_RETRY_AFTER_SECONDS",
+        new=0.1,
+    )
+
+    example_response_content = (
+        request.path.parent / "example_api_responses/getProject.json"
+    ).read_text()
+    expected_result = json.loads(example_response_content)
+
+    requests_mock.get(
+        "https://app.launchdarkly.com/api/v2/projects/test-project-key",
+        [
+            {"status_code": 429},
+            {"status_code": 200, "text": example_response_content},
+        ],
+    )
+
+    client = LaunchDarklyClient(token=token)
+
+    # When
+    result = client.get_project(project_key=project_key)
+
+    # Then
+    assert result == expected_result