+++ frontend/common/types/responses.ts
@@ -1,5 +1,3 @@
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-
 export type EdgePagedResponse<T> = PagedResponse<T> & {
   last_evaluated_key?: string
   pages?: (string | undefined)[]
@@ -246,7 +244,7 @@ export type GettingStartedTask = {
 export type Onboarding = {
   tools: {
     completed: boolean
-    integrations: string[]
+    selection: string[]
   }
   tasks: GettingStartedTask[]
 }
@@ -810,6 +808,62 @@ export type IdentityTrait = {
   trait_value: FlagsmithValue
 }
 
+export enum PipelineStatus {
+  DRAFT = 'DRAFT',
+  ACTIVE = 'ACTIVE',
+}
+
+export interface ReleasePipeline {
+  id: number
+  name: string
+  project: number
+  description: string
+  stages_count: number
+  published_at: string
+  published_by: number
+  features_count: number
+}
+
+export interface SingleReleasePipeline extends ReleasePipeline {
+  stages: PipelineStage[]
+  completed_features: number[]
+}
+
+export enum StageTriggerType {
+  ON_ENTER = 'ON_ENTER',
+  WAIT_FOR = 'WAIT_FOR',
+}
+
+export type StageTriggerBody = { wait_for?: string } | null
+
+export type StageTrigger = {
+  trigger_type: StageTriggerType
+  trigger_body: StageTriggerBody
+}
+
+export enum StageActionType {
+  TOGGLE_FEATURE = 'TOGGLE_FEATURE',
+  TOGGLE_FEATURE_FOR_SEGMENT = 'TOGGLE_FEATURE_FOR_SEGMENT',
+}
+
+export type StageActionBody = { enabled: boolean; segment_id?: number }
+export interface StageAction {
+  id: number
+  action_type: StageActionType
+  action_body: StageActionBody
+}
+
+export type PipelineStage = {
+  id: number
+  name: string
+  pipeline: number
+  environment: number
+  order: number
+  trigger: StageTrigger
+  actions: StageAction[]
+  features: number[]
+}
+
 export type Res = {
   segments: PagedResponse<Segment>
   segment: Segment
@@ -947,8 +1001,21 @@ export type Res = {
   conversionEvents: PagedResponse<ConversionEvent>
   splitTest: PagedResponse<SplitTestResult>
   onboardingSupportOptIn: { id: string }
+  environmentMetrics: {
+    metrics: {
+      value: number
+      description: string
+      name: string
+      entity: 'features' | 'identities' | 'segments' | 'workflows'
+      rank: number
+    }[]
+  }
   profile: User
   onboarding: {}
   userPermissions: UserPermissions
+  releasePipelines: PagedResponse<ReleasePipeline>
+  releasePipeline: SingleReleasePipeline
+  pipelineStages: PagedResponse<PipelineStage>
+  pipelineStage: PipelineStage
   // END OF TYPES
 }