+++ frontend/web/components/App.js
@@ -1,6 +1,5 @@
 import React, { Component, Fragment } from 'react'
-import { matchPath } from 'react-router'
-import { Link, withRouter } from 'react-router-dom'
+import { Link, withRouter, matchPath } from 'react-router-dom'
 import * as amplitude from '@amplitude/analytics-browser'
 import { plugin as engagementPlugin } from '@amplitude/engagement-browser'
 import { sessionReplayPlugin } from '@amplitude/plugin-session-replay-browser'
@@ -25,7 +24,7 @@ import GithubStar from './GithubStar'
 import Tooltip from './Tooltip'
 import classNames from 'classnames'
 import { apps, gitBranch, gitCompare, statsChart } from 'ionicons/icons'
-import NavSubLink from './NavSubLink'
+import NavSubLink from './navigation/NavSubLink'
 import SettingsIcon from './svg/SettingsIcon'
 import UsersIcon from './svg/UsersIcon'
 import BreadcrumbSeparator from './BreadcrumbSeparator'
@@ -42,10 +41,6 @@ import { getBuildVersion } from 'common/services/useBuildVersion'
 const App = class extends Component {
   static propTypes = {
     children: propTypes.element.isRequired,
-  }
-
-  static contextTypes = {
-    router: propTypes.object.isRequired,
   }
 
   state = {
@@ -132,16 +127,6 @@ const App = class extends Component {
     updateLastViewed()
   }
 
-  toggleDarkMode = () => {
-    const newValue = !Utils.getFlagsmithHasFeature('dark_mode')
-    flagsmith.setTrait('dark_mode', newValue)
-    if (newValue) {
-      document.body.classList.add('dark')
-    } else {
-      document.body.classList.remove('dark')
-    }
-  }
-
   componentDidUpdate(prevProps) {
     if (prevProps.location.pathname !== this.props.location.pathname) {
       const newProjectId = this.getProjectId(this.props)
@@ -200,7 +185,7 @@ const App = class extends Component {
 
     if (!AccountStore.getOrganisation() && !invite) {
       // If user has no organisation redirect to /create
-      this.context.router.history.replace(`/create${query}`)
+      this.props.history.replace(`/create${query}`)
       return
     }
 
@@ -215,11 +200,11 @@ const App = class extends Component {
     ) {
       if (redirect) {
         API.setRedirect('')
-        this.context.router.history.replace(redirect)
+        this.props.history.replace(redirect)
       } else {
         AsyncStorage.getItem('lastEnv').then((res) => {
           if (this.props.location.search.includes('github-redirect')) {
-            this.context.router.history.replace(
+            this.props.history.replace(
               `/github-setup${this.props.location.search}`,
             )
             return
@@ -230,7 +215,7 @@ const App = class extends Component {
               id: lastEnv.orgId,
             })
             if (!lastOrg) {
-              this.context.router.history.replace('/organisations')
+              this.props.history.replace('/organisations')
               return
             }
 
@@ -243,7 +228,7 @@ const App = class extends Component {
               AppActions.getOrganisation(lastOrg.id)
             }
 
-            this.context.router.history.replace(
+            this.props.history.replace(
               `/project/${lastEnv.projectId}/environment/${lastEnv.environmentId}/features`,
             )
             return
@@ -253,9 +238,9 @@ const App = class extends Component {
             Utils.getFlagsmithHasFeature('welcome_page') &&
             AccountStore.getUser()?.isGettingStarted
           ) {
-            this.context.router.history.replace('/getting-started')
+            this.props.history.replace('/getting-started')
           } else {
-            this.context.router.history.replace(Utils.getOrganisationHomePage())
+            this.props.history.replace(Utils.getOrganisationHomePage())
           }
         })
       }
@@ -282,7 +267,7 @@ const App = class extends Component {
     if (document.location.href.includes('saml?')) {
       return
     }
-    this.context.router.history.replace('/')
+    this.props.history.replace('/')
   }
 
   closeAnnouncement = (announcementId) => {
@@ -349,6 +334,12 @@ const App = class extends Component {
     }
     const isOrganisationSelect = document.location.pathname === '/organisations'
     const integrations = Object.keys(Utils.getIntegrationData())
+    const environmentMetricsEnabled = Utils.getFlagsmithHasFeature(
+      'environment_metrics',
+    )
+    const projectMetricsTooltipEnabled = Utils.getFlagsmithHasFeature(
+      'project_metrics_tooltip',
+    )
     return (
       <Provider store={getStore()}>
         <AccountProvider
@@ -439,7 +430,6 @@ const App = class extends Component {
                                     <div className='d-flex gap-1 ml-1 align-items-center'>
                                       <BreadcrumbSeparator
                                         projectId={projectId}
-                                        router={this.context.router}
                                         hideSlash={!activeProject}
                                         focus='organisation'
                                       >
@@ -466,7 +456,6 @@ const App = class extends Component {
                                       {!!activeProject && (
                                         <BreadcrumbSeparator
                                           projectId={projectId}
-                                          router={this.context.router}
                                           hideSlash
                                           focus='project'
                                         >
@@ -535,24 +524,6 @@ const App = class extends Component {
                                   <GithubStar />
 
                                   <Headway className='cursor-pointer' />
-                                  <Tooltip
-                                    place='bottom'
-                                    title={
-                                      <div className='dark-mode mt-0'>
-                                        <Switch
-                                          checked={Utils.getFlagsmithHasFeature(
-                                            'dark_mode',
-                                          )}
-                                          onChange={this.toggleDarkMode}
-                                          darkMode
-                                        />
-                                      </div>
-                                    }
-                                  >
-                                    {Utils.getFlagsmithHasFeature('dark_mode')
-                                      ? 'Light Mode'
-                                      : 'Dark Mode'}
-                                  </Tooltip>
                                 </Row>
                               </nav>
                             </React.Fragment>
@@ -617,6 +588,21 @@ const App = class extends Component {
                           >
                             Compare
                           </NavSubLink>
+                          {projectMetricsTooltipEnabled && (
+                            <NavSubLink
+                              icon={gitCompare}
+                              to=''
+                              id='reporting-link'
+                              disabled
+                              tooltip={
+                                Utils.getFlagsmithValue(
+                                  'project_metrics_tooltip',
+                                ) || 'Coming soon - fallback'
+                              }
+                            >
+                              Reporting
+                            </NavSubLink>
+                          )}
                           <Permission
                             level='project'
                             permission='ADMIN'
@@ -634,6 +620,27 @@ const App = class extends Component {
                               )
                             }
                           </Permission>
+                          {Utils.getFlagsmithHasFeature(
+                            'release_pipelines',
+                          ) && (
+                            <Permission
+                              level='project'
+                              permission='ADMIN'
+                              id={projectId}
+                            >
+                              {({ permission }) =>
+                                permission && (
+                                  <NavSubLink
+                                    icon={<Icon name='flash' />}
+                                    id='release-pipelines-link'
+                                    to={`/project/${projectId}/release-pipelines`}
+                                  >
+                                    Release Pipelines
+                                  </NavSubLink>
+                                )
+                              }
+                            </Permission>
+                          )}
                         </>
                       ) : (
                         !!AccountStore.getOrganisation() && (
@@ -712,7 +719,7 @@ const App = class extends Component {
                   {environmentId && !isCreateEnvironment ? (
                     <div className='d-flex'>
                       <HomeAside
-                        history={this.context.router.history}
+                        history={this.props.history}
                         environmentId={environmentId}
                         projectId={projectId}
                       />
@@ -735,6 +742,7 @@ const App = class extends Component {
 App.propTypes = {
   history: RequiredObject,
   location: RequiredObject,
+  match: RequiredObject,
 }
 
 export default withRouter(ConfigProvider(App))